Move SplitButton to its own folder
