+++ src/components/splitButton/SplitButton.js
@@ -1,8 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import Box, { pickBoxProps } from '../box';
-import Button from './Button';
-import ButtonGroup from './ButtonGroup';
+import Button, { ButtonGroup } from '../button';
 import { Menu } from '../menu';
 import Popover from '../popover';
 import { IconChevronDownSmallOutline } from '@teamleader/ui-icons';
